% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticmodelcreation.R
\name{logisticmodelcreation}
\alias{logisticmodelcreation}
\title{Logistic model creation}
\usage{
logisticmodelcreation(var)
}
\arguments{
\item{See}{the example}
}
\value{
you can download the model and all the performance
}
\description{
Give you ui to create logistic model.
}
\examples{
m<-data.frame()
x <- sapply(sapply(ls(), get), is.data.frame)
m <- data.frame(names(x)[(x == TRUE)])
logisticmodelcreation(m)
}
\author{
JIshnu
}
